requirements.txt
google-generativeai
discord-py-interactions